requirements.txt
streamlit
crawl4ai
openai
pydantic
python-dotenv
nest_asyncio
